Set control to use uint8 instead of double
This is because I noticed that the LSB steganography technique
consistently gave an even higher PSNR than the control, which does not
make sense at all. I then remembered that LSB steganography uses the
uint8 data type, whereas all the others have to use double because of
the transformations. By changing the control to use uint8 instead of
double for the image values, it maintains a higher PSNR, which is
correct.

+++ steganography/steganography_image_control.m
@@ -43,7 +43,7 @@ else
 end
 
 % Load image
-im = imload([dir_input, carrier_image_filename], use_greyscale);
+im = uint8(imload([dir_input, carrier_image_filename], use_greyscale));
 
 if use_greyscale
     imc = im;
@@ -69,7 +69,7 @@ imwrite(uint8(im_stego), [dir_output, output_image_filename], 'Quality', output_
 % ======
 
 % Read image and take chosen channel
-im_stego = imload([dir_output, output_image_filename], use_greyscale);
+im_stego = uint8(imload([dir_output, output_image_filename], use_greyscale));
 
 if use_greyscale
     imc_stego = im_stego;
